+++ website/src/main.tsx
@@ -6,6 +6,7 @@ import "./styles/NavbarLink.css";
 import "./styles/SearchBox.css";
 import "./styles/DarkModeButton.css";
 import "./styles/fonts.css";
+import "./styles/CustomFooter.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>